fix(image): int to size_t

+++ srcs/raytrace/get_image_ref_color.c
@@ -18,10 +18,10 @@ static t_map_idx	get_map_idx(t_tangetnt_map map, t_img img)
 
 	map.u *= (double)img.width * IMG_FREQUENCY;
 	map.v *= (double)img.height * IMG_FREQUENCY;
-	ret.row = ((int)map.u % img.width + img.width) % img.width;
-	ret.col = ((int)map.v % img.height + img.height) % img.height;
-	ret.idx = ((ret.col * img.width + ret.row) * PPM_RGB_UNIT) \
-				% (img.width * img.height * PPM_RGB_UNIT);
+	ret.row = ((int)map.u % (int)img.width + (int)img.width) % (int)img.width;
+	ret.col = ((int)map.v % (int)img.height + (int)img.height) % (int)img.height;
+	ret.idx = ((ret.col * (int)img.width + ret.row) * PPM_RGB_UNIT) \
+				% ((int)img.width * (int)img.height * PPM_RGB_UNIT);
 	return (ret);
 }
